server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = ""
authors = [
    {name = "Adrien Quennehen",email = "adrien.quennehen@descartesunderwriting.com"}
]
requires-python = ">=3.10,<4.0"
dependencies = [
    "djangorestframework (>=3.16.1,<4.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "drf-yasg (>=1.21.11,<2.0.0)",
    "drf-spectacular (>=0.28.0,<0.29.0)",
    "django-unfold (>=0.68.0,<0.69.0)",
    "pgvector (>=0.4.1,<0.5.0)",
    "django[gis] (>=5.2.7,<6.0.0)",
    "django-leaflet (>=0.32.0,<0.33.0)",
    "django-geojson (>=4.2.0,<5.0.0)",
    "djangorestframework-gis (>=1.2.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "black (>=25.9.0,<26.0.0)",
    "isort (>=7.0.0,<8.0.0)",
    "flake8 (>=7.3.0,<8.0.0)"
]

